work though this monster

+++ database/migrations/2025_08_06_100001_create_work_time_model_schedules_table.php
@@ -17,7 +17,7 @@ return new class() extends Migration
                 ->cascadeOnDelete();
 
             $table->unsignedInteger('week_number');
-            $table->unsignedInteger('weekday');
+            $table->unsignedTinyInteger('weekday');
             $table->time('start_time')->nullable();
             $table->time('end_time')->nullable();
             $table->decimal('break_minutes', 5, 2)->default(0);
